requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Marshmallow==0.15.0
marshmallow==3.21.1
marshmallow-sqlalchemy==0.29.0
GeoAlchemy2==0.14.4
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
python-dotenv==1.0.1
Shapely==2.0.4
pyshp==2.3.1
